turning FeedInfo service's download methods into a new FeedFetch service
This way it can also be used by the feed `fetch_and_return_feed_version` method, which previously used the Carrierwave gem's file downloading functionality. I suspect Carrierwave was loading the whole file into memory, before writing it to disk. This may have led to Sidekiq jobs bloating and temporary files being left on servers.

(hopefully) fixes #311
