+++ app/services/feed_info.rb
@@ -1,5 +1,3 @@
-require 'net/http'
-
 class FeedInfo
 
   CACHE_EXPIRATION = 1.hour
@@ -19,7 +17,7 @@ class FeedInfo
       @gtfs.load_graph
       yield self
     elsif @url
-      FeedInfo.download_to_tempfile(@url) do |path|
+      FeedFetch.download_to_tempfile(@url) do |path|
         @path = path
         @gtfs = GTFS::Source.build(@path, {strict: false})
         @gtfs.load_graph
@@ -58,44 +56,4 @@ class FeedInfo
     # done
     return [feed, operators]
   end
-
-  def self.download_to_tempfile(url, maxsize=nil)
-    fetch(url) do |response|
-      file = Tempfile.new('test.zip', Dir.tmpdir)
-      file.binmode
-      total = 0
-      begin
-        response.read_body do |chunk|
-          file.write(chunk)
-          total += chunk.size
-        end
-        raise IOError.new('Exceeds maximum file size') if (maxsize && total > maxsize)
-        file.close
-        yield file.path
-      ensure
-        file.close unless file.closed?
-        file.unlink
-      end
-    end
-  end
-
-  def self.fetch(url, limit=10, &block)
-    # http://ruby-doc.org/stdlib-2.2.3/libdoc/net/http/rdoc/Net/HTTP.html
-    # You should choose a better exception.
-    raise ArgumentError.new('Too many redirects') if limit == 0
-    url = URI.parse(url)
-    Net::HTTP.start(url.host, url.port) do |http|
-      http.request_get(url.request_uri) do |response|
-        case response
-        when Net::HTTPSuccess then
-          yield response
-        when Net::HTTPRedirection then
-          fetch(response['location'], limit-1, &block)
-        else
-          raise response.value
-        end
-      end
-    end
-  end
-
 end
