+++ spec/support/vcr_cassettes/feed_fetch.yml
@@ -21,11 +21,11 @@ http_interactions:
       Server:
       - nginx
       Date:
-      - Wed, 09 Dec 2015 00:39:40 GMT
+      - Fri, 22 Jan 2016 23:59:24 GMT
       Content-Type:
       - application/json
       Content-Length:
-      - '247'
+      - '244'
       Connection:
       - keep-alive
       Access-Control-Allow-Origin:
@@ -36,8 +36,8 @@ http_interactions:
       encoding: UTF-8
       string: "{\n  \"args\": {}, \n  \"headers\": {\n    \"Accept\": \"*/*\", \n
         \   \"Accept-Encoding\": \"gzip;q=1.0,deflate;q=0.6,identity;q=0.3\", \n    \"Host\":
-        \"httpbin.org\", \n    \"User-Agent\": \"Ruby\"\n  }, \n  \"origin\": \"172.125.87.109\",
+        \"httpbin.org\", \n    \"User-Agent\": \"Ruby\"\n  }, \n  \"origin\": \"99.51.77.14\",
         \n  \"url\": \"http://httpbin.org/get\"\n}\n"
     http_version: 
-  recorded_at: Wed, 09 Dec 2015 00:39:40 GMT
-recorded_with: VCR 3.0.0
+  recorded_at: Fri, 22 Jan 2016 23:59:24 GMT
+recorded_with: VCR 3.0.1
