+++ spec/support/vcr_cassettes/feed_fetch_download.yml
@@ -2,7 +2,7 @@
 http_interactions:
 - request:
     method: get
-    uri: http://httpbin.org/status/404
+    uri: http://httpbin.org/get
     body:
       encoding: US-ASCII
       string: ''
@@ -15,17 +15,17 @@ http_interactions:
       - Ruby
   response:
     status:
-      code: 404
-      message: NOT FOUND
+      code: 200
+      message: OK
     headers:
       Server:
       - nginx
       Date:
-      - Wed, 09 Dec 2015 00:30:39 GMT
+      - Fri, 22 Jan 2016 23:59:29 GMT
       Content-Type:
-      - text/html; charset=utf-8
+      - application/json
       Content-Length:
-      - '0'
+      - '244'
       Connection:
       - keep-alive
       Access-Control-Allow-Origin:
@@ -34,7 +34,10 @@ http_interactions:
       - 'true'
     body:
       encoding: UTF-8
-      string: ''
+      string: "{\n  \"args\": {}, \n  \"headers\": {\n    \"Accept\": \"*/*\", \n
+        \   \"Accept-Encoding\": \"gzip;q=1.0,deflate;q=0.6,identity;q=0.3\", \n    \"Host\":
+        \"httpbin.org\", \n    \"User-Agent\": \"Ruby\"\n  }, \n  \"origin\": \"99.51.77.14\",
+        \n  \"url\": \"http://httpbin.org/get\"\n}\n"
     http_version: 
-  recorded_at: Wed, 09 Dec 2015 00:30:38 GMT
-recorded_with: VCR 3.0.0
+  recorded_at: Fri, 22 Jan 2016 23:59:29 GMT
+recorded_with: VCR 3.0.1
