+++ spec/support/vcr_cassettes/feed_fetch_download_binary.yml
@@ -21,7 +21,7 @@ http_interactions:
       Server:
       - nginx
       Date:
-      - Wed, 09 Dec 2015 00:36:53 GMT
+      - Fri, 22 Jan 2016 23:59:30 GMT
       Content-Type:
       - application/octet-stream
       Transfer-Encoding:
@@ -59,5 +59,5 @@ http_interactions:
         T+aM+sWSQ690uGd+Bb0IrpTGSq81h1f6+DWQVPfslriuWurmVL8C0JDzXKBS
         yJn8ACQLIO3yevLRx78wjGzzLCuoKByBy5rBREht/bfyiQ==
     http_version: 
-  recorded_at: Wed, 09 Dec 2015 00:36:53 GMT
-recorded_with: VCR 3.0.0
+  recorded_at: Fri, 22 Jan 2016 23:59:30 GMT
+recorded_with: VCR 3.0.1
