+++ spec/support/vcr_cassettes/feed_fetch_redirect.yml
@@ -2,7 +2,7 @@
 http_interactions:
 - request:
     method: get
-    uri: http://httpbin.org/absolute-redirect/6
+    uri: http://httpbin.org/redirect-to?url=http://httpbin.org/get
     body:
       encoding: US-ASCII
       string: ''
@@ -21,31 +21,27 @@ http_interactions:
       Server:
       - nginx
       Date:
-      - Wed, 09 Dec 2015 00:27:11 GMT
+      - Fri, 22 Jan 2016 23:59:25 GMT
       Content-Type:
       - text/html; charset=utf-8
       Content-Length:
-      - '283'
+      - '0'
       Connection:
       - keep-alive
       Location:
-      - http://httpbin.org/absolute-redirect/5
+      - http://httpbin.org/get
       Access-Control-Allow-Origin:
       - "*"
       Access-Control-Allow-Credentials:
       - 'true'
     body:
       encoding: UTF-8
-      string: |-
-        <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 3.2 Final//EN">
-        <title>Redirecting...</title>
-        <h1>Redirecting...</h1>
-        <p>You should be redirected automatically to target URL: <a href="http://httpbin.org/absolute-redirect/5">http://httpbin.org/absolute-redirect/5</a>.  If not click the link.
+      string: ''
     http_version: 
-  recorded_at: Wed, 09 Dec 2015 00:27:11 GMT
+  recorded_at: Fri, 22 Jan 2016 23:59:26 GMT
 - request:
     method: get
-    uri: http://httpbin.org/absolute-redirect/5
+    uri: http://httpbin.org/get
     body:
       encoding: US-ASCII
       string: ''
@@ -58,32 +54,29 @@ http_interactions:
       - Ruby
   response:
     status:
-      code: 302
-      message: FOUND
+      code: 200
+      message: OK
     headers:
       Server:
       - nginx
       Date:
-      - Wed, 09 Dec 2015 00:27:11 GMT
+      - Fri, 22 Jan 2016 23:59:26 GMT
       Content-Type:
-      - text/html; charset=utf-8
+      - application/json
       Content-Length:
-      - '283'
+      - '244'
       Connection:
       - keep-alive
-      Location:
-      - http://httpbin.org/absolute-redirect/4
       Access-Control-Allow-Origin:
       - "*"
       Access-Control-Allow-Credentials:
       - 'true'
     body:
       encoding: UTF-8
-      string: |-
-        <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 3.2 Final//EN">
-        <title>Redirecting...</title>
-        <h1>Redirecting...</h1>
-        <p>You should be redirected automatically to target URL: <a href="http://httpbin.org/absolute-redirect/4">http://httpbin.org/absolute-redirect/4</a>.  If not click the link.
+      string: "{\n  \"args\": {}, \n  \"headers\": {\n    \"Accept\": \"*/*\", \n
+        \   \"Accept-Encoding\": \"gzip;q=1.0,deflate;q=0.6,identity;q=0.3\", \n    \"Host\":
+        \"httpbin.org\", \n    \"User-Agent\": \"Ruby\"\n  }, \n  \"origin\": \"99.51.77.14\",
+        \n  \"url\": \"http://httpbin.org/get\"\n}\n"
     http_version: 
-  recorded_at: Wed, 09 Dec 2015 00:27:11 GMT
-recorded_with: VCR 3.0.0
+  recorded_at: Fri, 22 Jan 2016 23:59:26 GMT
+recorded_with: VCR 3.0.1
