+++ spec/support/vcr_cassettes/feed_fetch_redirect_fail.yml
@@ -2,7 +2,7 @@
 http_interactions:
 - request:
     method: get
-    uri: http://httpbin.org/redirect-to?url=http://httpbin.org/get
+    uri: http://httpbin.org/absolute-redirect/6
     body:
       encoding: US-ASCII
       string: ''
@@ -21,27 +21,31 @@ http_interactions:
       Server:
       - nginx
       Date:
-      - Wed, 09 Dec 2015 00:08:48 GMT
+      - Fri, 22 Jan 2016 23:59:27 GMT
       Content-Type:
       - text/html; charset=utf-8
       Content-Length:
-      - '0'
+      - '283'
       Connection:
       - keep-alive
       Location:
-      - http://httpbin.org/get
+      - http://httpbin.org/absolute-redirect/5
       Access-Control-Allow-Origin:
       - "*"
       Access-Control-Allow-Credentials:
       - 'true'
     body:
       encoding: UTF-8
-      string: ''
+      string: |-
+        <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 3.2 Final//EN">
+        <title>Redirecting...</title>
+        <h1>Redirecting...</h1>
+        <p>You should be redirected automatically to target URL: <a href="http://httpbin.org/absolute-redirect/5">http://httpbin.org/absolute-redirect/5</a>.  If not click the link.
     http_version: 
-  recorded_at: Wed, 09 Dec 2015 00:08:47 GMT
+  recorded_at: Fri, 22 Jan 2016 23:59:27 GMT
 - request:
     method: get
-    uri: http://httpbin.org/get
+    uri: http://httpbin.org/absolute-redirect/5
     body:
       encoding: US-ASCII
       string: ''
@@ -54,29 +58,32 @@ http_interactions:
       - Ruby
   response:
     status:
-      code: 200
-      message: OK
+      code: 302
+      message: FOUND
     headers:
       Server:
       - nginx
       Date:
-      - Wed, 09 Dec 2015 00:08:48 GMT
+      - Fri, 22 Jan 2016 23:59:27 GMT
       Content-Type:
-      - application/json
+      - text/html; charset=utf-8
       Content-Length:
-      - '247'
+      - '283'
       Connection:
       - keep-alive
+      Location:
+      - http://httpbin.org/absolute-redirect/4
       Access-Control-Allow-Origin:
       - "*"
       Access-Control-Allow-Credentials:
       - 'true'
     body:
       encoding: UTF-8
-      string: "{\n  \"args\": {}, \n  \"headers\": {\n    \"Accept\": \"*/*\", \n
-        \   \"Accept-Encoding\": \"gzip;q=1.0,deflate;q=0.6,identity;q=0.3\", \n    \"Host\":
-        \"httpbin.org\", \n    \"User-Agent\": \"Ruby\"\n  }, \n  \"origin\": \"172.125.87.109\",
-        \n  \"url\": \"http://httpbin.org/get\"\n}\n"
+      string: |-
+        <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 3.2 Final//EN">
+        <title>Redirecting...</title>
+        <h1>Redirecting...</h1>
+        <p>You should be redirected automatically to target URL: <a href="http://httpbin.org/absolute-redirect/4">http://httpbin.org/absolute-redirect/4</a>.  If not click the link.
     http_version: 
-  recorded_at: Wed, 09 Dec 2015 00:08:48 GMT
-recorded_with: VCR 3.0.0
+  recorded_at: Fri, 22 Jan 2016 23:59:27 GMT
+recorded_with: VCR 3.0.1
